+++ services/backend/src/main/java/com/soprasteria/g4it/backend/apidigitalservice/modeldb/referential/DeviceTypeRef.java
@@ -61,9 +61,4 @@ public class DeviceTypeRef {
      * Source.
      */
     private String source;
-
-    /**
-     * Compatible Ecomind.
-     */
-    private boolean compatibleEcomind;
 }