+++ services/frontend/src/app/core/interfaces/digital-service.interfaces.ts
@@ -7,6 +7,7 @@
  */
 
 import { DropdownValue } from "./generic.interfaces";
+import { TaskRest } from "./inventory.interfaces";
 import { Note } from "./note.interface";
 
 export interface DigitalService {
@@ -23,6 +24,7 @@ export interface DigitalService {
     criteria?: string[];
     creator?: DigitalServiceUserInfo;
     isAi?: boolean;
+    tasks?: TaskRest[];
 }
 
 export interface DigitalServiceUserInfo {
@@ -144,6 +146,17 @@ export interface TerminalsType {
     code: string;
     value: string;
     lifespan: number;
+}
+
+export interface EcomindType {
+    code: string;
+    value: string;
+    lifespan: number;
+    defaultCpuCores: number;
+    defaultGpuCount: number;
+    defaultGpuMemory: number;
+    defaultRamSize: number;
+    defaultDatacenterPue: number;
 }
 
 export interface NetworkType {