+++ services/frontend/src/app/core/service/data/file-system-data.service.ts
@@ -6,20 +6,21 @@
  * French Ecological Ministery (https://gitlab-forge.din.developpement-durable.gouv.fr/pub/numeco/m4g/numecoeval)
  */
 import { HttpClient } from "@angular/common/http";
-import { Injectable } from "@angular/core";
+import { inject, Injectable } from "@angular/core";
+
 import { Observable } from "rxjs";
 import { Constants } from "src/constants";
 
-const endpoint = Constants.ENDPOINTS.inventories;
+const endpoint = Constants.ENDPOINTS.downloadReject;
 
 @Injectable({
     providedIn: "root",
 })
 export class FileSystemDataService {
-    constructor(private http: HttpClient) {}
+    private readonly http = inject(HttpClient);
 
-    downloadResultsFile(inventoryId: number, taskId: string): Observable<any> {
-        return this.http.get(`${endpoint}/${inventoryId}/output/${taskId}`, {
+    downloadResultsFile(taskId: string): Observable<any> {
+        return this.http.get(`${endpoint}/${taskId}`, {
             responseType: "blob",
             headers: { Accept: "application/zip" },
         });