+++ CHANGELOG.md
@@ -4,15 +4,19 @@
 
 ### Major Changes
 
+- 1357 | Integrate EcoMindAI in G4IT
+- 1204 | Import csv files for digital services
 - 1263 | Naming network entries in the Digital Service module – Network input
 - 1261 | Naming user groups in the Digital Service module – Terminales input
 - 740 | Visualize different cloud instances for the same type
 
 ### Minor Changes
 
+- 950 | Accessibility Corrections
+- 1320 | Mobile Screen Responsiveness for Top Header, Left side nav and Welcome page
 - 1255 | Add BoaviztAPI link to github
 - 1271 | Open Source - Be able to launch a calculation just after first installation
-- 1200 | [BE] New Bug on electricity consumption : total display different than sum compute manually with export file
+- 1200 | New Bug on electricity consumption : total display different than sum compute manually with export file
 - 1193 | DB cleanup:Remove the link between a user and a digital service
 - 1257 | Very High level of Digital Service calculated - Click in May
 - 1200 | New Bug on electricity consumption : total display different than sum compute manually with export file